% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_plot.R
\name{anomaly_plot}
\alias{anomaly_plot}
\title{Anomaly plot}
\usage{
anomaly_plot(object, data = NULL)
}
\arguments{
\item{object}{The fitted isoForest model object}

\item{data}{The data to be plotted.}
}
\value{
A ggplot object.
}
\description{
Plots the anomaly score heatmap for each point in the space.
}
\examples{
# Load the data
data("iris")
# Fit the anomaly detection model
model <- isoForest(iris[1:2])
# Plot the anomaly scores
anomaly_plot(model, iris[1:2])
}
