% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_nodes.r
\name{calculate_depth_per_tree}
\alias{calculate_depth_per_tree}
\title{Calculate the depth from leaf to root per tree.}
\usage{
calculate_depth_per_tree(df)
}
\arguments{
\item{df}{the tree info of ranger return by function 'treeInfo'.}
}
\value{
A dataframe with two columns: 'nodeID', 'depth'.
}
\description{
Depth of each terminal node of a single tree in a ranger model.
}
\examples{
rf <- ranger::ranger(Species ~ ., data = iris)
calculate_depth_per_tree(ranger::treeInfo(rf, 1))
}
