% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_nodes.r
\name{calculate_feature_counts}
\alias{calculate_feature_counts}
\title{Calculate the number of times a feature is used in terminal nodes.}
\usage{
calculate_feature_counts(model, obs_depth)
}
\arguments{
\item{model}{A ranger model}

\item{obs_depth}{A dataframe with columns: 'nodeID', 'treeID'.}
}
\value{
A dataframe with two columns: 'feature', 'count'.
}
\description{
The number of times a feature is used in terminal nodes of all
trees in a ranger model is returned as a dataframe with column names:
'feature', 'count'.
}
\examples{

result <- isoForest(iris, feature_contribution = TRUE)
head(result$feature_contributions)
}
