% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_nodes.r
\name{calculate_leaf_to_root_depth}
\alias{calculate_leaf_to_root_depth}
\title{Calculate the depth from leaf to root.}
\usage{
calculate_leaf_to_root_depth(model)
}
\arguments{
\item{model}{A ranger model}
}
\value{
A dataframe with three columns: 'nodeID', 'depth', 'treeID'.
}
\description{
Depth of each terminal node of all trees in a ranger model is
returned as a three column dataframe with column names: 'nodeID',
'depth', 'treeID'. Note that root node has the node_id = 0.
}
\examples{
rf <- ranger::ranger(Species ~ ., data = iris, num.trees = 100)
result <- calculate_leaf_to_root_depth(rf)
head(result)
}
