% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_contribution.R
\name{calculate_max_tree_depth}
\alias{calculate_max_tree_depth}
\title{Calculate maximum tree depth from tree structures}
\usage{
calculate_max_tree_depth(tree_infos)
}
\description{
Calculate maximum tree depth from tree structures
}
\keyword{internal}
