% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_contribution.R
\name{calculate_path_contributions_batch}
\alias{calculate_path_contributions_batch}
\title{Batch calculate path-based contributions (optimized)}
\usage{
calculate_path_contributions_batch(
  object,
  sample_ids,
  data,
  feature_names,
  max_trees
)
}
\description{
Batch calculate path-based contributions (optimized)
}
