% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_contribution.R
\name{calculate_permutation_contributions_batch}
\alias{calculate_permutation_contributions_batch}
\title{Batch calculate permutation importance (highly optimized)}
\usage{
calculate_permutation_contributions_batch(
  object,
  sample_ids,
  data,
  feature_names,
  n_permutations
)
}
\description{
Batch calculate permutation importance (highly optimized)
}
