% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{computeAnomaly}
\alias{computeAnomaly}
\title{Calculate the anomaly score}
\usage{
computeAnomaly(average_depth, sample_size)
}
\arguments{
\item{average_depth}{average depth of the terminal nodes}

\item{sample_size}{number of sample size}
}
\value{
anomaly score
}
\description{
Calculate the anomaly score
}
\examples{
\dontrun{
computeAnomaly(10, 256)
}
}
