% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{harmonic_approx}
\alias{harmonic_approx}
\title{Calculate the harmonic approximation}
\usage{
harmonic_approx(n)
}
\arguments{
\item{n}{number of sample size}
}
\value{
harmonic approximation
}
\description{
Calculate the harmonic approximation
}
\examples{
harmonic_approx(100)
}
