% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{pathLengthNormalizer}
\alias{pathLengthNormalizer}
\title{Calculate the normalization factor}
\usage{
pathLengthNormalizer(sample_size)
}
\arguments{
\item{sample_size}{number of sample size}
}
\value{
normalization factor
}
\description{
Calculate the normalization factor
}
\examples{
\dontrun{
pathLengthNormalizer(100)
}
}
