% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.isoForest}
\alias{predict.isoForest}
\title{Predict anomaly scores using an Isolation Forest model}
\usage{
\method{predict}{isoForest}(
  object,
  newdata,
  num.threads = NULL,
  type = "terminalNodes",
  seed = NULL,
  ...
)
}
\arguments{
\item{object}{an Isolation Forest model}

\item{newdata}{a data frame containing the new observations to be predicted}

\item{num.threads}{the number of threads to be used for the prediction (default is NULL, which means all available threads)}

\item{type}{the type of prediction to be made (default is "terminalNodes")}

\item{seed}{the seed to be used for the random number generator (default is NULL)}

\item{...}{additional arguments to be passed to the \code{predict} function}
}
\value{
a data frame containing the anomaly scores for the new observations
}
\description{
Predict anomaly scores using an Isolation Forest model
}
\examples{
# Load the example data
data(iris)
# Train an Isolation Forest model
model <- isoForest(iris[1:2])
# Predict anomaly scores for new observations
predictions <- predict(model, iris[1:2])
}
