% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{print.anomaly_threshold}
\alias{print.anomaly_threshold}
\title{Print method for anomaly_threshold objects}
\usage{
\method{print}{anomaly_threshold}(x, ...)
}
\arguments{
\item{x}{An anomaly_threshold object}

\item{...}{Additional arguments (not used)}
}
\description{
Print summary of anomaly threshold results
}
