% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_contribution.R
\name{print.feature_contribution}
\alias{print.feature_contribution}
\title{Print method for feature_contribution objects}
\usage{
\method{print}{feature_contribution}(x, top_n = 5, ...)
}
\arguments{
\item{x}{A feature_contribution object}

\item{top_n}{Number of top features to show per sample (default 5)}

\item{...}{Additional arguments (not used)}
}
\description{
Print feature contribution results
}
