% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_plot.R
\name{anomaly_plot}
\alias{anomaly_plot}
\title{Anomaly plot}
\usage{
anomaly_plot(
  object,
  data = NULL,
  plot_type = c("heatmap", "scatter"),
  contamination = 0.05
)
}
\arguments{
\item{object}{The fitted isoForest model object}

\item{data}{The data to be plotted.}

\item{plot_type}{The type of plot to be generated. Either "heatmap" or "scatter".}

\item{contamination}{The proportion of outliers in the data. Default is 0.05. This parameter is used to determine the threshold for anomaly scores.}
}
\value{
A ggplot object.
}
\description{
Plots the anomaly score heatmap for each point in the space.
}
\examples{
# Load the data
data("iris")
# Fit the anomaly detection model
model <- isoForest(iris[1:2])
# Plot the anomaly scores
anomaly_plot(model, iris[1:2],plot_type="scatter")
anomaly_plot(model, iris[1:2],plot_type="heatmap")
}
